channel-header component and new info route
