+++ app/styles/components/_list.scss
@@ -2,6 +2,13 @@
 .List {
 	list-style: none;
 	padding-left: 0;
+
+	background-color: $slightgray;
+	border-top: 1px solid hsl(0, 0%, 85%);
+	border-left: 1px solid hsl(0, 0%, 80%);
+	border-right: 1px solid hsl(0, 0%, 80%);
+	border-radius: $border-radius;
+	box-shadow: $box-shadow;
 }
 
 .List-header {
