+++ tests/integration/components/channel-header/component-test.js
@@ -1,0 +1,25 @@
+import { moduleForComponent, test } from 'ember-qunit';
+import hbs from 'htmlbars-inline-precompile';
+
+moduleForComponent('channel-header', 'Integration | Component | channel header', {
+  integration: true
+});
+
+test('it renders', function(assert) {
+
+  // Set any properties with this.set('myProperty', 'value');
+  // Handle any actions with this.on('myAction', function(val) { ... });
+
+  this.render(hbs`{{channel-header}}`);
+
+  assert.equal(this.$().text().trim(), '');
+
+  // Template block usage:
+  this.render(hbs`
+    {{#channel-header}}
+      template block text
+    {{/channel-header}}
+  `);
+
+  assert.equal(this.$().text().trim(), 'template block text');
+});
